requirements.txt
EbookLib
requests
gTTS
beautifulsoup4
google-generativeai
python-dotenv
